change type of instruction_count

+++ src/types.rs
@@ -26,4 +26,4 @@ pub type Fs2 = u8;
 pub type Fs1 = u8;
 pub type Funct2 = u8;
 pub type Fd = u8;
-pub type InstructionCount = u64;
+pub type InstructionCount = u128;
